pyproject.toml
[tool.poetry]
name = "repl_python3_beta_socketio-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gevent = "^21.1.2"
gevent-websocket = "^0.10.1"
python-socketio = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
